+++ test/nimble_totp_test.exs
@@ -60,13 +60,16 @@ defmodule NimbleTOTPTest do
 
     test "generate different codes in different periods (default is 30s)" do
       secret = NimbleTOTP.secret()
-      time1 = to_unix(~N[2020-04-08 17:49:59Z])
-      time2 = to_unix(~N[2020-04-08 17:50:00Z])
-      time3 = to_unix(~N[2020-04-08 17:50:30Z])
+      time1 = ~N[2020-04-08 17:49:59Z]
+      time2 = ~N[2020-04-08 17:50:00Z]
+      time3 = ~N[2020-04-08 17:50:30Z]
 
       code1 = NimbleTOTP.verification_code(secret, time: time1)
+      assert code1 == NimbleTOTP.verification_code(secret, time: to_unix(time1))
       code2 = NimbleTOTP.verification_code(secret, time: time2)
+      assert code2 == NimbleTOTP.verification_code(secret, time: to_unix(time2))
       code3 = NimbleTOTP.verification_code(secret, time: time3)
+      assert code3 == NimbleTOTP.verification_code(secret, time: to_unix(time3))
 
       codes = [code1, code2, code3]
 
@@ -87,12 +90,50 @@ defmodule NimbleTOTPTest do
   describe "valid?/2" do
     test "returns true if it matches the verification code" do
       time = System.os_time(:second)
+      date_time = DateTime.from_unix!(time, :second)
+      naive_date_time = DateTime.to_naive(date_time)
+
+      for _ <- 1..1000 do
+        secret = NimbleTOTP.secret()
+
+        code = NimbleTOTP.verification_code(secret, time: time)
+        assert code == NimbleTOTP.verification_code(secret, time: date_time)
+        assert code == NimbleTOTP.verification_code(secret, time: naive_date_time)
+
+        assert NimbleTOTP.valid?(secret, code, time: time)
+        assert NimbleTOTP.valid?(secret, code, time: date_time)
+        assert NimbleTOTP.valid?(secret, code, time: naive_date_time)
+
+        refute NimbleTOTP.valid?(secret, "abcdef", time: time)
+        refute NimbleTOTP.valid?(secret, "abcdef", time: date_time)
+        refute NimbleTOTP.valid?(secret, "abcdef", time: naive_date_time)
+      end
+    end
+
+    test "rejects reused verification codes" do
+      time = System.os_time(:second)
+      next_time = (Integer.floor_div(time, 30) + 1) * 30
 
       for _ <- 1..1000 do
         secret = NimbleTOTP.secret()
         code = NimbleTOTP.verification_code(secret, time: time)
+        next_code = NimbleTOTP.verification_code(secret, time: next_time)
         assert NimbleTOTP.valid?(secret, code, time: time)
         refute NimbleTOTP.valid?(secret, "abcdef", time: time)
+
+        # Rejects all invalid codes regardless of the since option
+        refute NimbleTOTP.valid?(secret, "abcdef", time: time, since: time)
+        refute NimbleTOTP.valid?(secret, "abcdef", time: time, since: next_time)
+        refute NimbleTOTP.valid?(secret, "abcdef", time: time, since: nil)
+
+        # If since is nil (e.g. first time the code is entered)
+        assert NimbleTOTP.valid?(secret, code, time: time, since: nil)
+
+        # If the code was just entered
+        refute NimbleTOTP.valid?(secret, code, time: time, since: time)
+
+        # If the next code is entered in the last time-step window
+        assert NimbleTOTP.valid?(secret, next_code, time: next_time, since: time)
       end
     end